add tab in profile
