+++ pages/profile/index.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import styled from "styled-components";
-import ProfileArea from "../../src/components/profile/MyPage/ProfileArea";
-import StudyArea from "../../src/components/profile/MyPage/StudyArea";
+import ProfileArea from "../../src/components/profile/profileArea";
+import StudyArea from "../../src/components/profile/studyArea";
 
 
 // const ProfileArea = styled.div`
@@ -9,7 +9,7 @@ import StudyArea from "../../src/components/profile/MyPage/StudyArea";
 // `;
 
 
-export default function MyPage():JSX.Element {
+export default function Index():JSX.Element {
     return (
         <div>
             <ProfileArea/>
