+++ pages/profile/rewards.tsx
